update fixtures, tables

+++ src/Controller/Admin/UserController.php
@@ -21,6 +21,8 @@ class UserController extends AbstractController
      */
     public function index(UserRepository $userRepository): Response
     {
+        $this->denyAccessUnlessGranted('list', new User());
+
         return $this->render('admin/user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -32,6 +34,9 @@ class UserController extends AbstractController
     public function new(Request $request, UserPasswordEncoderInterface $encoder): Response
     {
         $user = new User();
+
+        $this->denyAccessUnlessGranted('create', $user);
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -55,6 +60,8 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
+        $this->denyAccessUnlessGranted('show', $user);
+
         return $this->render('admin/user/show.html.twig', [
             'user' => $user,
         ]);
@@ -65,6 +72,8 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('edit', $user);
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -85,6 +94,8 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('delete', $user);
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
